trading_bot_project/requirements.txt
pandas>=1.4
numpy>=1.20
requests>=2.25
PyYAML>=6.0
schedule>=1.1
python-dotenv>=1.0
ta>=0.11.0
python-telegram-bot>=20.0
slack_sdk>=3.21
matplotlib>=3.7